# src/utils/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VUMI - Komiku Rest API",
    "version": "1.0.0",
    "description": "API Backend yang dioptimasi untuk Aplikasi Comic Reader Vumi. Dikembangkan dengan Node.js dan Express oleh Kibbz Skibidi.",
    "contact": {
      "name": "Kibbz Skibidi",
      "url": "https://sakibdev.my.id"
    }
  },
  "servers": [
    {
      "url": "https://api-vumi-scrap.vercel.app",
      "description": "Production Server (Vercel)"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/vumi/baca-chapter/{slug}/{chapter}": {
      "get": {
        "tags": ["Comic Details"],
        "summary": "Konten gambar per chapter",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "lookism"
          },
          {
            "name": "chapter",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "1"
          }
        ],
        "responses": {
          "200": { "description": "Gambar chapter berhasil dimuat" },
          "404": { "description": "Chapter tidak ditemukan" }
        }
      }
    },
...
  }
}